ported to Grails-3.0.1
