--- application.properties
@@ -1,4 +1,0 @@
-#Grails Metadata file
-#Sat Mar 22 10:54:17 CDT 2014
-app.grails.version=2.3.7
-app.name=clojure
