--- grails-app/conf/BuildConfig.groovy
@@ -1,42 +1,0 @@
-grails.servlet.version = "3.0" // Change depending on target container compliance (2.5 or 3.0)
-grails.project.class.dir = "target/classes"
-grails.project.test.class.dir = "target/test-classes"
-grails.project.test.reports.dir = "target/test-reports"
-grails.project.work.dir = "target/work"
-grails.project.target.level = 1.6
-grails.project.source.level = 1.6
-//grails.project.war.file = "target/${appName}-${appVersion}.war"
-
-grails.project.dependency.resolver = "maven" // or ivy
-grails.project.dependency.resolution = {
-    // inherit Grails' default dependencies
-    inherits("global") {
-        // specify dependency exclusions here; for example, uncomment this to disable ehcache:
-        // excludes 'ehcache'
-    }
-    log "error" // log level of Ivy resolver, either 'error', 'warn', 'info', 'debug' or 'verbose'
-    checksums true // Whether to verify checksums on resolve
-    legacyResolve false // whether to do a secondary resolve on plugin installation, not advised and here for backwards compatibility
-
-    repositories {
-        inherits true // Whether to inherit repository definitions from plugins
-
-        mavenRepo 'http://build.clojure.org/releases'
-        grailsPlugins()
-        grailsHome()
-        mavenLocal()
-        grailsCentral()
-        mavenCentral()
-    }
-
-    dependencies {
-         runtime 'org.clojure:clojure:1.5.1'
-    }
-
-    plugins {
-        // plugins for the build system only
-        build ":tomcat:7.0.50.1"
-        build ":release:3.0.1"
-        compile ":functional-test:2.0.RC1"
-    }
-}
