+++ grails-app/views/demo/plain.gsp
@@ -1,0 +1,11 @@
+<%@ page contentType="text/html;charset=UTF-8" %>
+<html>
+<head>
+    <meta name="layout" content="main"/>
+    <title>${text}</title>
+</head>
+
+<body>
+    <h1>${text}</h1>
+</body>
+</html>
