--- ivy.xml
@@ -1,28 +1,0 @@
-<ivy-module version="2.0">
-    <info organisation="org.example" module="clojure"/>
-    <configurations defaultconfmapping="build->default;compile->compile(*),master(*);test,runtime->runtime(*),master(*)">
-        <conf name="build"/>
-        <conf name="compile"/>
-        <conf name="test" extends="compile"/>
-        <conf name="runtime" extends="compile"/>
-    </configurations>
-    <dependencies>
-        <dependency org="org.grails" name="grails-bootstrap" rev="1.1.1" conf="build"/>
-        <dependency org="org.grails" name="grails-scripts" rev="1.1.1" conf="build"/>
-        <dependency org="org.grails" name="grails-gorm" rev="1.1.1" conf="compile"/>
-        <dependency org="org.grails" name="grails-web" rev="1.1.1" conf="compile"/>
-        <dependency org="org.grails" name="grails-test" rev="1.1.1" conf="test"/>
-        <dependency org="org.slf4j" name="slf4j-log4j12" rev="1.5.5" conf="runtime"/>
-        <dependency org="opensymphony" name="oscache" rev="2.4" conf="runtime">
-            <exclude org="javax.jms" module="jms" name="*" type="*" ext="*" conf="" matcher="exact"/>
-            <exclude org="commons-logging" module="commons-logging" name="*" type="*" ext="*" conf="" matcher="exact"/>
-            <exclude org="javax.servlet" module="servlet-api" name="*" type="*" ext="*" conf="" matcher="exact"/>
-        </dependency>
-        <dependency org="hsqldb" name="hsqldb" rev="1.8.0.5" conf="runtime"/>
-		<dependency org="net.sf.ehcache" name="ehcache" rev="1.5.0" conf="runtime"/>
-        <!--
-        <dependency org="mysql" name="mysql-connector-java" rev="5.1.6" conf="runtime"/>
-        <dependency org="postgresql" name="postgresql" rev="8.3-603.jdbc3" conf="runtime"/>
-        -->
-    </dependencies>
-</ivy-module>
